Add interaction to contact section
In order to center the marker properly in the cutout, I made the map a
fixed height and set the center. Once the map loads, a resize is
triggered and the new center is fetched. What remains to be taken care
of is window resizing; the center needs to be recalculated (at least
when the height of the window is changed).

TODO:

- Obtain a background with the cutout, position it properly
- Might need a proper close/go back button or something later
- Get the marker icon and put it in
- Figure out what the header should do when the map gets brought to the
  forefront